+++ sprout/views/projects_id_metadata_id/step_confirmation.py
@@ -2,8 +2,8 @@ from django.http import HttpRequest, HttpResponse, HttpResponseRedirect
 from django.shortcuts import render
 
 
-def projects_id_metadata_create_4(
+def step_confirmation(
     request: HttpRequest, table_id: int
 ) -> HttpResponse | HttpResponseRedirect:
     """Renders page final page when creating metadata table."""
-    return render(request, "projects-id-metadata-create.html")
+    return render(request, "projects_id_metadata_id/create.html")
